apps/api/api_common/pyproject.toml
# Copyright © Advanced Micro Devices, Inc., or its affiliates.
#
# SPDX-License-Identifier: MIT

[build-system]
requires = ["setuptools"]

[project]
name = "api_common"
description = "Common functionality and schema for AIRM and AIWB APIs"
version = "0.1.0"
readme = "README.md"
requires-python = ">=3.13"
classifiers = ["Programming Language :: Python"]
dependencies = [
    "fastapi>=0.115.12",
    "httpx>=0.28.1",
    "loguru>=0.7.3",
    "pydantic>=2.11.3",
    "python-dotenv>=1.1.0",
    "pyyaml>=6.0.2",
    "sqlalchemy>=2.0.0",
    "python-keycloak>=4.6.1",
]

[tool.setuptools]
include-package-data = true

[tool.setuptools.packages.find]
where = ["."]
include = ["api_common*"]

[tool.pytest.ini_options]
doctest_optionflags = "NORMALIZE_WHITESPACE IGNORE_EXCEPTION_DETAIL"
log_cli = "0"
log_cli_level = "INFO"

[tool.mypy]
strict = false
files = ["**/*.py"]
ignore_missing_imports = true
pretty = true
show_column_numbers = true

[dependency-groups]
dev = [
    "coverage>=7.8.0",
    "docker>=7.1.0",
    "pre-commit>=4.2.0",
    "pytest>=8.3.5",
    "pytest-asyncio>=0.24.0",
    "pytest-docker>=3.2.1",
    "pytest-env>=1.1.5",
    "pytest-httpx>=0.35.0",
    "responses>=0.25.7",
    "uvicorn>=0.34.0",
]
